+++ src/custom-components/maps/MainMap.vue
@@ -45,7 +45,7 @@
                         <VsAlert
                             v-if="noResults || (props.alertText && noResults === false)"
                             class="mt-075 mb-150"
-                            id="vs-map__no-results-alert"
+                            id="vs-map__alert"
                             size="small"
                         >
                             <template v-if="noResults">
@@ -883,8 +883,6 @@ async function searchByCategory({
 
     const searchId = currentSearchId;
 
-    noResults.value = false;
-
     googleMapStore.filterUsesCount += 1;
 
     currentSearch.value = 'nearby';
@@ -935,7 +933,6 @@ async function searchByCategory({
 async function searchByText() {
     resetMap();
     resetCategories();
-    noResults.value = false;
 
     currentSearchId += 1;
 
@@ -996,6 +993,7 @@ async function addMarkers(searchId) {
         noResults.value = true;
         searchRequest.value.style.display = 'none';
     } else {
+        noResults.value = false;
         searchRequest.value.style.display = 'block';
     }
 
@@ -1208,9 +1206,7 @@ function checkFirstInteraction(interactionType) {
 function getVisibleMarkerCount() {
     const bounds = gMap.getBounds();
 
-    if (!bounds) {
-        return 0;
-    }
+    if (!bounds) return 0;
 
     let visibleCount = 0;